Moving status to its own app for less dependency. Going to have to really rework the apps. This is not close to a working or tested build. Don't use. Fixing up the models and prepping them. Finally back into the nTracking monster. I'll update the rig a bit more. Still not around the db to actually build and test. Here we go.
