+++ noctis/nStatus/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class NstatusConfig(AppConfig):
+    name = 'nStatus'
