+++ requirements.txt
@@ -1,5 +1,5 @@
 Brlapi==0.6.1
-Django==1.9.1
+Django==1.9.5
 Mako==0.9.1
 MarkupSafe==0.18
 apturl==0.5.2ubuntu4
@@ -10,6 +10,7 @@ colorama==0.2.5
 command-not-found==0.3
 defer==1.0.6
 django-json-rpc==0.7.2
+django-simple-history==1.8.1
 feedparser==5.1.3
 friends==0.1
 html5lib==0.999
